requirements.txt

jupyter
pdf2image

transformers
requests
Pillow
accelerate
opencv-python
pdf2image
openai-whisper
setuptools-rust
pyaudio
packaging
ninja

pydub
librosa
scikit-image
scikit-video
pandas